refactor: Mock-up Data, Repository 분리
