+++ app/src/main/java/com/jukco/waitforme/data/repository/AuthProvider.kt
@@ -2,7 +2,6 @@ package com.jukco.waitforme.data.repository
 
 import androidx.compose.runtime.Composable
 import com.jukco.waitforme.ui.sign.social.AuthUiProvider
-import com.jukco.waitforme.ui.sign.social.MockAuthUiProvider
 
 interface AuthProvider {
     @Composable
